gpssim: Add docbook-style documentation to structures + functions
I felt that the current code could deal with a lot more comments,
particularly with somebody not familiar with all the very-few-letter
mnemonics of th pseudorange equation.  It is my hope that by adding
these comments, somebody with a basic understanding of GPS can read
through the code more easily.